requirements.txt
backoff==2.2.1
beautifulsoup4==4.13.3
botocore==1.36.26
certifi==2025.1.31
charset-normalizer==3.4.1
coloredlogs==15.0.1
deprecation==2.1.0
filelock==3.17.0
flatbuffers==25.2.10
fsspec==2025.2.0
gdown==5.2.0
gitdb==4.0.12
h5py==3.13.0
huggingface-hub==0.29.1
humanfriendly==10.0
idna==3.10
imageio==2.37.0
Jinja2==3.1.5
jmespath==1.0.1
lazy_loader==0.4
MarkupSafe==3.0.2
mpmath==1.3.0
networkx==3.4.2
numpy==1.24.4
onnx==1.16.1
onnxruntime==1.18.1
opencv-python==4.9.0.80
packaging==24.2
pandas==2.2.3
pillow==11.1.0
prettytable==3.14.0
protobuf==3.20.3
PySocks==1.7.1
python-dateutil==2.9.0.post0
pytz==2025.1
PyYAML==6.0.2
qai-hub==0.24.0
requests==2.32.3
requests-toolbelt==1.0.0
ruamel.yaml==0.18.10
ruamel.yaml.clib==0.2.12
s3transfer==0.10.4
schema==0.7.7
scikit-image==0.23.2
scipy==1.15.2
semver==3.0.4
six==1.17.0
smmap==5.0.2
soupsieve==2.6
sympy==1.13.1
tabulate==0.9.0
tifffile==2025.2.18
torch==2.6.0
torchvision==0.21.0
tqdm==4.67.1
typing_extensions==4.12.2
tzdata==2025.1
urllib3==2.3.0
wcwidth==0.2.13
